requirements.txt
# requirements.txt

# Core FastAPI framework and server
fastapi
uvicorn[standard]

# Database ORM and PostgreSQL driver
sqlalchemy
psycopg2-binary

# Password hashing
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# JWT token handling
python-jose[cryptography]

# For loading .env files
python-dotenv

# For email validation in schemas
pydantic[email]

#newly added for sdm
httpx
itsdangerous
python-multipart

apscheduler
python-dateutil
google-auth
google-auth-oauthlib
google-api-python-client

langchain
langchain_core
langchain_openai
beautifulsoup4
pandas
selenium
webdriver-manager
pypdf2
requests-html
selenium-stealth